feat(carousel-card-section-offset): add print styles to Carousel and Card Section Offset (#7649)
### Related Ticket(s)

Github: https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/7414

### Description

"This request originally stemmed from discussions with the case study team and the Afterburner experience guild team as we collaborated on the new case study article template. We wanted the Carbon for ibm.com components to be able to be assembled together in a way that when the user chooses to print the web page, it is visually aligned with what the components should look like. Specifically for case study team, this will allow the team to create custom PDFs for the more important case studies and rely on the print-friendly view for all other case studies. This use case seems applicable not just for the case study team, but for all other pages that uses the Carbon components (e.g. Learn, Solution)."

### Changelog

**New**

- Print styles for the Carousel and Card Offset Section