fix(masthead): account for platform name in active item logic (#7890)
### Related Ticket(s)
Resolves https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/7794.

### Description

There are currently two ways menu items can be set to "active" and receive a blue underline:
1. Manually: Set the `selected-menu-item` property on the `<dds-masthead-composite>` element
2. Automatically: A menu item has a URL that is identical the the current page's URL.

Additionally, there is a concept of a "platform" supported by the optional `platform` and `platformUrl` properties. When these properties are set, a bolded link appears first in the masthead's L1 nav, and the tricky part is that this link also has logic for whether or not it is set to active. This logic was aware of the manually set active menu items, but it did not account for the automatically set ones.

<!-- Bug reproduced here: https://codesandbox.io/s/festive-wescoff-cvl2h?file=/index.html -->

### Changelog

**Changed**

- Adds `_hasAutoSelectedItems` internal property to track whether automatically selected menu items are present or not.
- Uses `_hasAutoSelectedItems` internal property as a factor when determining if platform link should receive active styles.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->