chore(deploy): replace deployment of canary/next envs to Github Pages (#8013)
### Related Ticket(s)

No related issue

### Description

With the deprecation of Cloud Foundry on IBM Cloud, plus finding a performant solution for running nightly Browserstack e2e tests, it was important to find a solution for hosting all of our lower testing environments. 

This PR introduces a move to auto deploy all of our test environments for Carbon for IBM.com to Github Pages. It will generate into the following folders:

- canary/react
- canary/react-rtl
- canary/react-experimental
- canary/web-components
- canary/web-components-rtl
- canary/web-components-experimental
- canary/web-components-react
- staging/react
- staging/react-rtl
- staging/web-components
- staging/web-components-rtl
- services
- utilities

This was tested on my fork and seems to be working fine now:
- https://jeffchew.github.io/carbon-for-ibm-dotcom/canary/react
- https://jeffchew.github.io/carbon-for-ibm-dotcom/canary/web-components
- see other folders above

### Changelog

**Changed**

- deploy scripts changed from IBM Cloud to Github Pages
- shortened `STORYBOOK_` environment variables to be consistent and not as lengthy
- Added scripts specific for rtl and experimental storybook builds in React and Web Components
- Added `DefinePlugin` for more explicit `STORYBOOK_USE_RTL` variable usage for static deployments in React

**Removed**

- All manifest files