fix(card): announce content change on keyboard focus (#8031)
### Related Ticket(s)

#7899
#7900

### Description

This PR updates the pictogram card content area so that screen readers will announce content changes on focus. The `aria-describedby` attribute will also point to the card copy container element. Not sure if we want to refactor this component to contain an anchor as #7900 mentions but if so I may pull that out into a separate PR

Confirm that when focusing on the card, the screen reader announces the changed text. Also make sure that the card announces its role as an interactive group and mentions that the card can be interacted with

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->