feat(masthead): hide and show upon scroll on mobile (#7982)
### Related Ticket(s)
#6046

### Description
This PR adds a feature to the React masthead that hides it upon scroll on mobile if the page contains only the Masthead L0 and Table of Contents component. This also used to be a feature on the Web Components version, but due to some logic updates a few months ago the functionality remained unable to run. This PR restores this feature on WC.

This can be tested in the default Dotcom Shell story in both React and WC environments in a mobile breakpoint.

### Changelog

**New**

- added hide/show masthead scroll functionality to react

**Changed**

- restored hide/show masthead scroll functionality in web components

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->