fix(table-of-contents): fix erroneous heading check (#8059)
### Related Ticket(s)
Fixes #7663 

### Description
Adjusts logic from slotChange event handler to not rely on the event object. For some reason, the `slotChange` event was firing twice, once with the slot properly referenced in `event.target` and once without. This led to the `_hasHeading` property getting out-of-sync. This PR uses `this.querySelectorAll` to find the nodes that _should_ be assigned to the `heading` slot.

### Changelog

**Changed**

- fixed issue where TOC heading wouldn't display after resize