fix(dotcom-shell): stop intersection logic if modal is open (#8083)
### Related Ticket(s)
#8078

### Description
There was some previous logic to avoid this bug from happening, but some check may have slipped through that made it reoccur. Instead of adding a check to every conditional, this PR only adds a simple check if the locale modal is open, and stops any code that might make the browser "scroll" in the `handleIntersection` function and would make the modal disappear. 

This needs to be tested in an actual phone (or Browserstack).

### Changelog
**Changed**

- changed location of the conditional check for the local modal being open to stop it from disappearing

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->