fix(e2e): mock ibm-common.js in e2e tests (#8147)
### Related Ticket(s)

Refs https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/7898

### Description

The footer tests were still failing for the footer in Browserstack automate. This is an attempt to approach the fix in a different way, and instead to block the cookie preferences (ibm-common.js) from loading entirely. 

This also introduces the approach for globally mocking network calls rather than doing it in a per-test approach.

### Changelog

**New**

- `ibm-common.js` fixture in e2e tests

**Changed**

- moved cypress fixtures in global support index.js