fix(content-section): update child elements alignment to the top (#8182)
### Related Ticket(s)

#7804 

### Description

When a child element is added to content-section we need to remove the top of either padding or margin, so the element aligns to the content-section heading. 

Note: there are multiple style differences from the design. This PR though is just trying to get the headings aligned. When trying to fix the other styles (ie text and spacing), it was causing a lot of percy diffs. Thinking we should tackle one at a time. Will create an issue with the style discrepancies. 

### Changelog

**New**

- wc: slotchange
- react useEffect
- added cta-section styles, so that those headings still align


<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->