feat(search-with-typeahead): add scoped search support (#8096)
### Related Ticket(s)
#7627

### Description
This PR changes the SearchWithTypeahead API slightly so it's able to retrieve scoped results from different categories within IBM, such as PartnerWorld or Developer as examples. This can be tested under the `Masthead - with scoped results` in WC Storybook. Change the scope by clicking on the filter select button while the results are displayed to see how they change between scopes.

### Changelog

**New**

- scoped search in search with typeahead

**Changed**

- added scope to `SearchWithTypeahead` API query if requested

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->