fix(filter-panel): update heading when knob changes (#8362)
### Related Ticket(s)

Resolves https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/7441

### Description

The `<dds-filter-panel-composite>`'s heading appears in multiple places, each of which is essentially copied from a source `<dds-filter-panel-heading>` in the `heading` slot. The problem was that changes to this source were not being reflected in all the places that copy it (i.e. the modal button, the modal header, and the desktop header).

The changes in this PR result in changes to the story's "heading" knob updating the heading everywhere it appears.

### Changelog

**New**

- `DDSFilterPanelHeading` emits an event when its slotted content changes.
- `DDSFilterPanelComposite` propagates changes to source header to modal button text, modal header text, and desktop header text.

**Changed**

- Changes to the "heading" knob are reflected in the component.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->