chore(gh-actions): add timeout to e2e tests (#8378)
### Description

This adds a timeout to the `e2e-tests` GitHub jobs for both React and web components. These jobs can hang for hours and clog the pipeline. At a quick glance, these jobs can finish as quickly as 14 minutes to as long as 35 minutes (or longer, depending on the amount of jobs in the queue). Initial timeout value is set to 60, which can be adjusted as needed.

### Changelog

**New**

- add `timeout-minutes` to end-to-end Github jobs

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->