chore(github-actions): activate e2e integration tests in Github Actions (#8396)
### Related Ticket(s)

No related issue

### Description

This is a new workflow that will run the e2e integration tests on push to `main`, `release/*`, and nightly. With this in place, I'm removing the PR CSB e2e testing (but will still generate the codesandbox examples for now, unless we want to remove that too? or just run it with a label?).

This will hopefully free up the Jenkins jobs tremendously.

### Changelog

**New**

- `e2e-integration.yml`