fix(masthead): aligned L1 sticky with wc (#8398)
### Related Ticket(s)
#8032

### Description
The previous Masthead L1 sticky code was only making so that the masthead disappeared immediately the moment the user started scrolling down without the table of contents pushing it up due to the way it was being activated. It also introduced a weird bug in iOS where the Masthead could simply disappear scrolling up. 

With this PR, the method used for sticky behavior aligns with the way it's being done in WC, so it's smoother all around.

### Changelog

**Changed**

- using new method to define sticky behavior with L1

**Removed**

- removed sticky booleans and states inside masthead 

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->