fix(filter-panel): return focus to selected elements (#8492)
### Related Ticket(s)

Resolves https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/8307

### Description

Filter panel elements (e.g. checkbox, select item) were losing focus when clicked/selected. This was happening because the whole filter panel is re-rendered whenever an element is clicked, and as far as the browser is concerned, the element that had focus disappeared. As a result, it was dropping focus back to the window.

To fix this, the composite now tracks a matching selector for any element gets selected. It uses this selector to find that element and reapply focus to it after any re-renders.

### Changelog

**New**

- Adds `_focusElement` state property to `DDSFilterPanelComposite` elements. This property tracks which filter element was last selected and is used to reapply focus to that element after re-renders.

**Changed**

- Renames `DDSFilterPanelComposite` events and event listeners to better match those events' purposes. 

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->