fix(masthead): fix search combobox `aria-expanded` value (#8662)
### Related Ticket(s)

#8473

### Description

Currently the search combobox `aria-expanded` value is based on whether or not the suggestions list is open. This is incorrect, as the value should come from the search input (combobox) itself. This would cause the value to be `false` unless the suggestions list was visible.

Note: making this change in React introduced a new accessibility violation, which I believe is a false positive. The combobox popup (suggestions list) has no focus or child elements, and `aria-activedescendant` is only set when the list populates and a selection is highlighted. I'll follow up with the a11y team about this.

![image](https://user-images.githubusercontent.com/909118/160870480-d5030219-831e-47a1-87c7-1e735a426bab.png)


### Changelog

**Changed**

- change `aria-expanded` value to be based on search input state, not suggestions list

**Removed**

- unused React Storybook DOM element

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->