fix(masthead): using better selector (#8737)
### Related Ticket(s)
None.

### Description
The React Canary tests were failing recently due to the fact that the previous selector that was using the `data-autoid` over the menu class was selecting the `Let's Create` menu that was implemented kind of recently. For some reason it was only failing in Chrome100 , Big Sur but it was okay in the rest of the browsers. Hopefully this PR makes it so it will always attempt to select the first menu that contains a submenu as the tests intended.

### Changelog

**New**

- {{new thing}}

**Changed**

-`data-autoid="dds--masthead-default__l0-nav0"]` to `.bx--header__submenu` to only specify the first menu that is selectable

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->