fix(tabs-extended & table-of-contents): ensure DOM updates are handled (#8767)
### Related Ticket(s)

Resolves  #8751

### Description

#### DDSTabsExtended
There were cases where certain boolean attributes were being parsed as strings (e.g. `"false"`). Since non-empty strings are truthy, the value `"false"` was treated as `true`, and this caused any new tabs injected into the DOM to be  disabled.

This has been fixed by declaring to Lit Element that these attributes should be reflected as booleans. This tells Lit Element to convert the strings `"true"` and `"false"` to the booleans `true` and `false`, respectively.

#### DDSTableOfContents
The `slotchange` event that we typically use to catch updates to slotted DOM nodes has a limitation. From [the MDN docs](https://developer.mozilla.org/en-US/docs/Web/API/HTMLSlotElement/slotchange_event):

> Note: the slotchange event doesn't fire if the children of a slotted node change — only if you change (e.g. add or delete) the actual nodes themselves.

This means the event will not fire if one of the immediate descendent DOM nodes are merely updated (e.g. have an attribute added/removed), nor will it fire if any of their children are added, removed, or modified.

Instead, we have to implement a MutationObserver to track changes to those nodes and rebuild the table of contents accordingly.

### Changelog

**Added**
- Added a MutationObserver to `DDSTableOfContents` components that rebuilds the table of contents when it sees changes to descendants of the default slot's children.

**Changed**

- Declare `boolean` type for `DDSTab`'s boolean reactive properties.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->