feat(attribute-transpose): introduced alt attribute transposing (#8967)
### Related Ticket(s)
#8436 

### Description
The feature request consisted on ensuring that the `button` element that targets the video to play in the `dds-video-player-container` component to also transpose its `href` attribute along with the `data-*` attributes. 

Since the current video player doesn't include an `href` attribute, this PR also slightly modifies the component by dynamically creating the appropriate mediacenter URL that contains the video in question and insert it as the component's `href` attribute. In case a custom `href` is manually set, it will override the mediacenter link but will still be transposed to the `button` element as requested.

In other words
```
<dds-video-player-container video-id='example123'>
```

Would be reflected upon load to the following:

```
<dds-video-player-container href='https://mediacenter.ibm.com/id/example123' video-id='example123'>
   // within shadow root
   <button href='https://mediacenter.ibm.com/id/example123'>
```

The changes also make it possible for any future attributes to be added as well to generalize the transpose method to not only take in the `data-*` attributes by any others desired as well.

### Changelog

**New**

- added additional optional argument to the `transposeAttributes` method in `StableSelector` mixin to take in optional additional attributes to transpose

**Changed**
- video player container now sets an `href` that defaults to the mediacenter URL containing the video based on the video-id provided


<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->