fix(lightbox-with-carousel): fix focus wrapping logic & carousel focus movement (#9004)
### Related Ticket(s)

Fixes #8994 

### Description

This PR addresses two accessibility problems.

## Fixing focus wrapping in modals with complex content

This PR refactors the focus-wrapping of the DDSExpressiveModal. Prior to this PR, the expressive modal uses the internal [_focusableElements](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/blob/main/packages/web-components/src/components/expressive-modal/expressive-modal.ts#L129) property to hold onto a reference of all focusable elements so that we can wrap focus to the first or last focusable element within the modal. This property is [populated when the component is first updated](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/blob/main/packages/web-components/src/components/expressive-modal/expressive-modal.ts#L414-L417) with all descendants and shadowdom elements that match the `selectorTabbable` string

```js
this._focusableElements = [
  ...Array.from((this.shadowRoot?.querySelectorAll(selectorCloseButton) as NodeListOf<HTMLElement>) || []),
  ...Array.from(this.querySelectorAll(selectorTabbableForModal) as NodeListOf<HTMLElement>),
];
```
The problem with this approach, is that it's only querying the slotted elements of the modal and the modal's own shadowroot. Any focusable elements within slotted shadowroots aren't accounted for, and thus aren't a part of the modal's focus wrapping mechanism. This PR addresses that by changing the private, internal `_focusableElements` property into a getter method, and adding a public `focusableElements` getter method to the modal and any component that might be placed within a modal (right now it's just the carousel).

In this way, each component is responsible for keeping track of it's own focusable elements, and the modal is responsible for asking for those when it needs them. When the modal deems it necessary to wrap focus, it asks the elements it knows about to report their focusable elements, removes any duplicates, sorts the array of items, and then passes that polished list to the `tryFocusElems()` function. This approach _does_ cause re-querying each time the getter is run, but that's necessary for the carousel as the navigation buttons get disabled based on the status of the carousel.

## Fixing focus handling on carousel items that are off-screen

This PR adds some logic to handle focus movements in carousels. Prior to this fix, focusable elements in carousel slides off-screen could still gain focus and move the carousel slides. This creates a difference in how the page is percieved between sighted mouse users and visually-impaired keyboard users. Now, when the carousel's `start` property changes and slides are moved, the slides that are off-screen are given the [`inert`](https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/inert) attribute. This attribute is a boolean value that, when present, makes the browser "ignore" user input events for the element, including focus events and events from assistive technologies. This prevents focus from moving to slides that are off-screen, keeping parity between mouse and keyboard navigation. 

Right now, the inert fix would only impact a subset of users, as it's not fully implemented across all browsers, but it's a progressive enhancement feature that won't impact users that _don't_ have support for it at all. If desired, there's also a [polyfill](https://github.com/WICG/inert) that could bring inert functionality to most, if not all, browsers we currently support

### Changelog

**Changed**

- Updated focus-wrap logic within DDSExpressiveModal
- Updated carousel with inert attribute on hidden carousel items
- Added focusableItems getter to DDSCarousel for DDSExpressiveModal compatibility