fix(card-group): apply empty card styles in extended components (#9168)
### Related Ticket(s)

Resolves #9147

### Description

The CSS that hides the extra cards added to make the grid work was scoped to only ever apply to empty cards within `<dds-card-group>` elements, and those styles could not be extended (e.g. in Sass) since they were bundled up in the `DDSCardGroup` class. This made it impossible to hide the extra cards in components that want to extend from this class, such as our `CAEMCardGroup`, without copying and pasting CSS from the C4IBM source code.

This PR makes changes that should have no visual impact on the DDS components but will allow downstream inheritors to receive styles that are important to this component's appearance.

### Changelog

**Changed**

- Moves CSS styles that handle hiding the empty cards into the `ibmdotcom-styles` package so they can be extended if necessary.
- Removes unnecessary specificity from CSS rules so that empty cards in extending components are targeted by styles that hide them.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->