fix(pricing-table): accessibility QA updates (#9205)
### Related Ticket(s)

Resolves #9159, #9162 

### Description

#### Ensure header cells are fully visible regardless of zoom (#9159)
The pricing table header is capable of being taller than the viewport. When this is the case, users are not able to view the entirety of the header's contents before it transitions to its sticky state, which hides much of the cell content. This scenario is uncommon at the default viewport zoom level, but it becomes progressively more likely as the browser is zoomed in. Our Storybook examples exhibit the problem at around 400% zoom.

The best solution I could come up with is to disable the sticky functionality when the header is taller than the viewport. This seems better than the alternative, which is delaying the sticky transition until the user reaches the bottom of the header row. In any scenario where the browser window is not tall enough to display the entirety of the header, the value of having a sticky header is vastly diminished. For example, here's what a sticky header at 400% zoom looks like on my 13" Macbook:

<img width="1436" alt="Screen Shot 2022-08-08 at 5 02 24 PM" src="https://user-images.githubusercontent.com/12532727/183514108-48c71cc2-dbf8-4168-995e-83878b2a19f0.png">

Compare that to the version of the component where sticky header has been disabled:

<img width="1436" alt="Screen Shot 2022-08-08 at 5 03 31 PM" src="https://user-images.githubusercontent.com/12532727/183514218-37979e25-7c11-471c-bafc-b92e95728d39.png">

IMHO the non-sticky version seems much more usable. Note that this will affect short browser windows at default zoom levels:

With sticky (current behavior): 

https://user-images.githubusercontent.com/12532727/183514780-f53ef46b-708f-48be-85a3-41220daf1c02.mov

Without sticky (changes introduced in this PR): 

https://user-images.githubusercontent.com/12532727/183514841-7461550b-d55e-46e5-b8bb-a31111640aad.mov

#### Ensure Storybook contents are fully visible on scroll (#9162)

Scrolling down in a Pricing Table story causes the sticky header transition to start, but since there is no other content, it immediately bounces back into its initial position. I fixed this by adding an arbitrary 400px of space beneath the Pricing Tables in Storybook so that users can fully scroll down.

#### Account for new cells in sticky transitions

Adding new columns in Storybook then scrolling down revealed the new columns' header cells were not animating closed during the sticky transition. I've updated the logic to be aware of the new cells and include them in the animation.

#### Prevent unnecessary intersection observer resets

While debugging, I noticed the intersection observers were being reset during each frame of the sticky header animation. I added logic to prevent this unnecessary computation that could cause unexpected behaviors.

### Changelog

**Changed**

- `DDSPricingTable`s' sticky header behavior will no longer trigger when the table header is taller than the viewport.
- Prevents resetting `DDSPricingTable` intersection observers unnecessarily.
- Adds space at the bottom of Pricing Table stories so users can scroll down without sticky header pulling their viewport back up.
- `DDSPricingTableHeaderCell`s added after initial render are animated during sticky header transitions.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "test: e2e": Codesandbox examples and e2e integration tests -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->